Safari-609.1.11.1/LayoutTests/fast/parser/append-child-followed-by-document-write-expected.txt

ALERT: document.firstChild.outerHTML: <a></a>
document.firstChild.nextSibling.outerHTML: <html><head></head><body><b id="b"></b></body></html>
